requirements.txt
apache-airflow==2.3.4
apache-airflow-providers-microsoft-azure==4.2.0
pendulum==2.1.2